tsconfig refactor
