+++ apps/desktop/package.json
@@ -15,7 +15,6 @@
     "@sd/core": "workspace:*",
     "@sd/interface": "workspace:*",
     "@sd/ui": "workspace:*",
-    "@sd/vite": "workspace:*",
     "@tauri-apps/api": "^1.0.0-rc.3",
     "react": "^18.0.0",
     "react-dom": "^18.0.0"
@@ -37,7 +36,6 @@
     "typescript": "^4.6.3",
     "vite": "^2.9.5",
     "vite-plugin-filter-replace": "^0.1.9",
-    "vite-plugin-react-svg": "^0.2.0",
-    "vite-tsconfig-paths": "^3.4.1"
+    "vite-plugin-svgr": "^1.1.0"
   }
 }
