+++ apps/desktop/tsconfig.json
@@ -1,28 +1,5 @@
 {
-  "compilerOptions": {
-    "target": "ESNext",
-    "lib": ["DOM", "DOM.Iterable", "ESNext"],
-    "allowJs": false,
-    "skipLibCheck": false,
-    "esModuleInterop": false,
-    "allowSyntheticDefaultImports": true,
-    "strict": true,
-    "forceConsistentCasingInFileNames": true,
-    "module": "ESNext",
-    "moduleResolution": "Node",
-    "resolveJsonModule": true,
-    "isolatedModules": true,
-    "noEmit": true,
-    "jsx": "react",
-    "outDir": "dist",
-    "baseUrl": "./",
-    "types": ["vite-plugin-svgr/client"],
-    "paths": {
-      "@sd/interface": ["../../packages/interface/src/index.ts"],
-      "@sd/ui": ["../../packages/ui/src/index.ts"],
-      "@sd/client": ["../../packages/client/src/index.ts"],
-      "@sd/vite": ["../../packages/vite/src/index.ts"]
-    }
-  },
+  "extends": "../../packages/config/interface.tsconfig.json",
+  "compilerOptions": {},
   "include": ["src"]
 }
