+++ apps/desktop/vite.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
 import { name, version } from './package.json';
-import svg from '@sd/vite';
+import svg from "vite-plugin-svgr"
 
 // https://vitejs.dev/config/
 export default defineConfig({
@@ -13,7 +13,6 @@ export default defineConfig({
     react({
       jsxRuntime: 'classic'
     }),
-    //@ts-ignore
     svg({ svgrOptions: { icon: true } })
   ],
   root: 'src',
