+++ apps/landing/src/pages/changelog.tsx
@@ -1,6 +1,6 @@
 import Markdown from '../components/Markdown';
 import React from 'react';
-import { ReactComponent as Content } from '../../../../docs/changelog/index.md';
+import { ReactComponent as Content } from '~/docs/changelog/index.md';
 import { Helmet } from 'react-helmet';
 
 function Page() {
