+++ apps/landing/src/pages/docs/architecture/distributed-data-sync.tsx
@@ -1,6 +1,6 @@
 import Markdown from '../../../components/Markdown';
 import React from 'react';
-import { ReactComponent as Content } from '../../../../../../docs/architecture/distributed-data-sync.md';
+import { ReactComponent as Content } from '~/docs/architecture/distributed-data-sync.md';
 import { Helmet } from 'react-helmet';
 
 function Page() {
