+++ apps/landing/src/pages/faq.tsx
@@ -1,6 +1,6 @@
 import Markdown from '../components/Markdown';
 import React from 'react';
-import { ReactComponent as Content } from '../../../../docs/product/faq.md';
+import { ReactComponent as Content } from '~/docs/product/faq.md';
 import { Helmet } from 'react-helmet';
 
 function Page() {
