+++ apps/landing/vite.config.ts
@@ -1,22 +1,26 @@
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
-import svg from '@sd/vite';
 import pages from 'vite-plugin-pages';
 import md, { Mode } from 'vite-plugin-markdown';
-import generateSitemap from 'vite-plugin-pages-sitemap';
+import svg from 'vite-plugin-svgr';
 
 // https://vitejs.dev/config/
 export default defineConfig({
   // @ts-ignore
   plugins: [
     react(),
-    svg({ svgrOptions: { icon: true } }),
     pages({
       dirs: 'src/pages'
       // onRoutesGenerated: (routes) => generateSitemap({ routes })
     }),
+    svg(),
     md({ mode: [Mode.REACT] })
   ],
+  resolve: {
+    alias: {
+      '~/docs': __dirname + '../../../docs'
+    }
+  },
   server: {
     port: 8003
   },
