+++ apps/web/package.json
@@ -13,11 +13,8 @@
     "@sd/core": "*",
     "@sd/interface": "*",
     "@sd/ui": "*",
-    "@sd/vite": "workspace:*",
     "react": "^18.0.0",
-    "react-dom": "^18.0.0",
-    "react-tsparticles": "^2.0.6",
-    "tsparticles": "^2.0.6"
+    "react-dom": "^18.0.0"
   },
   "devDependencies": {
     "@types/react": "^18.0.6",
@@ -28,6 +25,7 @@
     "tailwind": "^4.0.0",
     "typescript": "^4.6.3",
     "vite": "^2.9.5",
-    "vite-plugin-svgr": "^1.1.0"
+    "vite-plugin-svgr": "^1.1.0",
+    "vite-plugin-tsconfig-paths": "^1.0.5"
   }
 }
