--- packages/client/.eslintrc.js
@@ -1,1 +1,0 @@
-module.exports = require('scripts/eslint-preset');
