+++ packages/config/interface.tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": "./base.tsconfig.json",
+  "compilerOptions": {
+    "types": ["vite-plugin-svgr/client"]
+  }
+}
