--- packages/config/tsconfig/react-library.json
@@ -1,11 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/tsconfig",
-  "display": "React Library",
-  "extends": "./base.json",
-  "compilerOptions": {
-    "lib": ["ES2015"],
-    "module": "ESNext",
-    "target": "ES6",
-    "jsx": "react-jsx"
-  }
-}
