+++ packages/interface/tsconfig.json
@@ -1,28 +1,7 @@
 {
+  "extends": "../config/interface.tsconfig.json",
   "compilerOptions": {
-    "target": "ESNext",
-    "lib": ["DOM", "DOM.Iterable", "ESNext"],
-    "allowJs": false,
-    "skipLibCheck": true,
-    "esModuleInterop": false,
-    "allowSyntheticDefaultImports": true,
-    "strict": true,
-    "forceConsistentCasingInFileNames": true,
-    "module": "ESNext",
-    "moduleResolution": "Node",
-    "resolveJsonModule": true,
-    "noEmit": true,
-    "jsx": "react",
-    "outDir": "dist",
-    "declaration": true,
-    "declarationMap": true,
-    "types": ["vite-plugin-svgr/client"],
-    "baseUrl": "./",
-    "paths": {
-      "@sd/ui": ["../ui/src/index.ts"],
-      "@sd/client": ["../client/src/index.ts"],
-      "@sd/core": ["../../core/index.ts"],
-    }
+    "outDir": "./dist"
   },
-  "include": ["src", "scripts"]
+  "include": ["src"]
 }
