+++ packages/ui/tsconfig.json
@@ -1,7 +1,7 @@
 {
-  "extends": "../config/tsconfig/react-library.json",
+  "extends": "../config/base.tsconfig.json",
   "include": ["src"],
   "compilerOptions": {
-    "outDir": "dist"
+    "outDir": "./dist"
   }
 }
