--- packages/vite/.prettierrc
@@ -1,12 +1,0 @@
-{
-  "useTabs": false,
-  "printWidth": 100,
-  "tabWidth": 2,
-  "singleQuote": true,
-  "trailingComma": "none",
-  "jsxBracketSameLine": false,
-  "semi": true,
-  "quoteProps": "consistent",
-  "importOrderSeparation": true,
-  "importOrderSortSpecifiers": true
-}
