--- packages/vite/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "@sd/vite",
-  "version": "1.0.0",
-  "license": "MIT",
-  "main": "dist/index.js",
-  "scripts": {
-    "dev": "tsc --watch",
-    "build": "tsc",
-    "prepare": "npm run build"
-  },
-  "dependencies": {
-    "@svgr/core": "^6.2.1",
-    "typescript": "^4.6.3",
-    "vite": "^2.9.6"
-  }
-}
